requirements/base.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
beautifulsoup4==4.13.4
certifi==2025.7.14
colorlog==6.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.4.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.1
wheel==0.45.1
